BFPSLGPlugin Source 1.0.11
With at altered implementation of the heuristics, that now only uses one
tile for each respective corner (instead of choosing a tile from a set).
Changed the matching values as well, as part of another guideline (blue
edges should tend to red and red should tend to blue, in a nutshell).